refactor: wrap state transitions in DB transactions and add domain events
- Wrap submit/reject/complete/cancel in executeInTransaction for SalesReturn, DeliveryOrder
- Add PaymentSent event dispatch for bill payments
- Add guard clause blocking sales returns from draft/cancelled invoices
- Remove paid_amount from Invoice $fillable (payments via service only)
- Add isEarlyPaymentDiscountAvailable() to Invoice model
- Add cache key registry to QuotationStatisticsService
- Simplify PaymentController error handling

+++ app/Domain/Sales/Quotations/TaxCalculator.php
@@ -8,7 +8,7 @@ use App\Models\Sales\QuotationItem;
 
 class TaxCalculator
 {
-    public static function calculateFromItems(iterable $items, int $taxRate = 11): int
+    public static function calculateFromItems(iterable $items, float $taxRate = 11): int
     {
         $subtotal = 0;
 
@@ -23,7 +23,7 @@ class TaxCalculator
         return self::calculateFromSubtotal($subtotal, $taxRate);
     }
 
-    public static function calculateFromSubtotal(int $subtotal, int $taxRate): int
+    public static function calculateFromSubtotal(int $subtotal, float $taxRate): int
     {
         return (int) round($subtotal * ($taxRate / 100));
     }
